[fix] fixed api coupon service, added test for fincoupon, fixed Dockerfile
