+++ .gitignore
@@ -1,2 +1,2 @@
-
-.idea/
+debug/
+.vscode/
