+++ coupon_service/Dockerfile
@@ -1,0 +1,27 @@
+# Start from the latest golang base image
+FROM golang:1.19-alpine AS builder
+
+# Set the Current Working Directory inside the container
+WORKDIR /app
+
+# Copy everything from the current directory to the Working Directory inside the container
+COPY . .
+
+# Build the Go app
+RUN go build -o main ./cmd/coupon_service/main.go
+
+# Start a new stage from scratch
+FROM alpine:latest  
+
+RUN apk --no-cache add ca-certificates
+
+WORKDIR /root/
+
+# Copy the pre-built binary file from the previous stage
+COPY --from=builder /app/main .
+
+# Expose port 8080 to the outside world
+EXPOSE 8080
+
+# Command to run the executable
+CMD ["./main"]
