+++ coupon_service/internal/api/entity/Coupons.go
@@ -1,0 +1,7 @@
+package entity
+
+type Coupon struct {
+	Discount       int
+	Code           string
+	MinBasketValue int
+}
