+++ coupon_service/internal/api/entity/applicationRequest.go
@@ -1,0 +1,8 @@
+package entity
+
+import "coupon_service/internal/service/entity"
+
+type ApplicationRequest struct {
+	Code   string
+	Basket entity.Basket
+}
