+++ coupon_service/internal/api/entity/couponRequest.go
@@ -1,0 +1,5 @@
+package entity
+
+type CouponRequest struct {
+	Codes []string
+}
