+++ coupon_service/internal/service/entity/basket.go
@@ -1,0 +1,16 @@
+package entity
+
+/*
+	Not needed
+
+import (
+
+	_ "github.com/gin-gonic/gin"
+
+)
+*/
+type Basket struct {
+	Value                 int
+	AppliedDiscount       int
+	ApplicationSuccessful bool
+}
