--- review/Dockerfile
@@ -1,14 +1,0 @@
-# build stage
-FROM golang:latest AS builder
-
-RUN apk add git gcc libc-dev
-
-WORKDIR /go/src/coupon-service
-
-COPY . .
-
-WORKDIR /go/src/coupon-service/cmd/export
-RUN go build -o main .
-
-ENTRYPOINT ./main
-EXPOSE 8080
