--- review/cmd/coupon_service/main.go
@@ -1,25 +1,0 @@
-package main
-
-import (
-	"coupon_service/internal/api"
-	"coupon_service/internal/config"
-	"coupon_service/internal/repository/memdb"
-	"coupon_service/internal/service"
-	"fmt"
-	"time"
-)
-
-var (
-	cfg  = config.New()
-	repo = memdb.New()
-)
-
-func main() {
-	svc := service.New(repo)
-	本 := api.New(cfg.API, svc)
-	本.Start()
-	fmt.Println("Starting Coupon service server")
-	<-time.After(1 * time.Hour * 24 * 365)
-	fmt.Println("Coupon service server alive for a year, closing")
-	本.Close()
-}
