--- review/internal/api/entity/Coupons.go
@@ -1,7 +1,0 @@
-package entity
-
-type Coupon struct {
-	Discount       int
-	Code           string
-	MinBasketValue int
-}
