--- review/internal/api/entity/applicationRequest.go
@@ -1,8 +1,0 @@
-package entity
-
-import "coupon_service/internal/service/entity"
-
-type ApplicationRequest struct {
-	Code   string
-	Basket entity.Basket
-}
