--- review/internal/api/entity/couponRequest.go
@@ -1,5 +1,0 @@
-package entity
-
-type CouponRequest struct {
-	Codes []string
-}
