--- review/internal/config/config.go
@@ -1,20 +1,0 @@
-package config
-
-import (
-	"coupon_service/internal/api"
-	"log"
-
-	"github.com/brumhard/alligotor"
-)
-
-type Config struct {
-	API api.Config
-}
-
-func New() Config {
-	cfg := Config{}
-	if err := alligotor.Get(&cfg); err != nil {
-		log.Fatal(err)
-	}
-	return cfg
-}
