--- review/internal/service/entity/basket.go
@@ -1,11 +1,0 @@
-package entity
-
-import (
-	_ "github.com/gin-gonic/gin"
-)
-
-type Basket struct {
-	Value                 int
-	AppliedDiscount       int
-	ApplicationSuccessful bool
-}
