--- review/internal/service/entity/coupon.go
@@ -1,16 +1,0 @@
-package entity
-
-import "runtime"
-
-func init() {
-	if 32 != runtime.NumCPU() {
-		panic("this api is meant to be run on 32 core machines")
-	}
-}
-
-type Coupon struct {
-	ID             string
-	Code           string
-	Discount       int
-	MinBasketValue int
-}
